+++ crates/model/src/instruments/any.rs
@@ -17,10 +17,12 @@ use enum_dispatch::enum_dispatch;
 use serde::{Deserialize, Serialize};
 
 use super::{
-    Instrument, betting::BettingInstrument, binary_option::BinaryOption,
-    crypto_future::CryptoFuture, crypto_option::CryptoOption, crypto_perpetual::CryptoPerpetual,
-    currency_pair::CurrencyPair, equity::Equity, futures_contract::FuturesContract,
-    futures_spread::FuturesSpread, option_contract::OptionContract, option_spread::OptionSpread,
+    Instrument, betting::BettingInstrument, binary_option::BinaryOption, cfd::Cfd,
+    commodity::Commodity, crypto_future::CryptoFuture, crypto_option::CryptoOption,
+    crypto_perpetual::CryptoPerpetual, currency_pair::CurrencyPair, equity::Equity,
+    futures_contract::FuturesContract, futures_spread::FuturesSpread,
+    index_instrument::IndexInstrument, option_contract::OptionContract,
+    option_spread::OptionSpread,
 };
 use crate::types::{Price, Quantity};
 
@@ -29,6 +31,8 @@ use crate::types::{Price, Quantity};
 pub enum InstrumentAny {
     Betting(BettingInstrument),
     BinaryOption(BinaryOption),
+    Cfd(Cfd),
+    Commodity(Commodity),
     CryptoFuture(CryptoFuture),
     CryptoOption(CryptoOption),
     CryptoPerpetual(CryptoPerpetual),
@@ -36,6 +40,7 @@ pub enum InstrumentAny {
     Equity(Equity),
     FuturesContract(FuturesContract),
     FuturesSpread(FuturesSpread),
+    IndexInstrument(IndexInstrument),
     OptionContract(OptionContract),
     OptionSpread(OptionSpread),
 }
@@ -47,6 +52,8 @@ impl InstrumentAny {
         match self {
             Self::Betting(inst) => inst.calculate_base_quantity(quantity, last_px),
             Self::BinaryOption(inst) => inst.calculate_base_quantity(quantity, last_px),
+            Self::Cfd(inst) => inst.calculate_base_quantity(quantity, last_px),
+            Self::Commodity(inst) => inst.calculate_base_quantity(quantity, last_px),
             Self::CryptoFuture(inst) => inst.calculate_base_quantity(quantity, last_px),
             Self::CryptoOption(inst) => inst.calculate_base_quantity(quantity, last_px),
             Self::CryptoPerpetual(inst) => inst.calculate_base_quantity(quantity, last_px),
@@ -54,6 +61,7 @@ impl InstrumentAny {
             Self::Equity(inst) => inst.calculate_base_quantity(quantity, last_px),
             Self::FuturesContract(inst) => inst.calculate_base_quantity(quantity, last_px),
             Self::FuturesSpread(inst) => inst.calculate_base_quantity(quantity, last_px),
+            Self::IndexInstrument(inst) => inst.calculate_base_quantity(quantity, last_px),
             Self::OptionContract(inst) => inst.calculate_base_quantity(quantity, last_px),
             Self::OptionSpread(inst) => inst.calculate_base_quantity(quantity, last_px),
         }