+++ nautilus_trader/test_kit/rust/instruments_pyo3.py
@@ -22,6 +22,8 @@ from nautilus_trader.core import nautilus_pyo3
 from nautilus_trader.core.nautilus_pyo3 import AssetClass
 from nautilus_trader.core.nautilus_pyo3 import BettingInstrument
 from nautilus_trader.core.nautilus_pyo3 import BinaryOption
+from nautilus_trader.core.nautilus_pyo3 import Cfd
+from nautilus_trader.core.nautilus_pyo3 import Commodity
 from nautilus_trader.core.nautilus_pyo3 import CryptoFuture
 from nautilus_trader.core.nautilus_pyo3 import CryptoOption
 from nautilus_trader.core.nautilus_pyo3 import CryptoPerpetual
@@ -30,6 +32,7 @@ from nautilus_trader.core.nautilus_pyo3 import CurrencyPair
 from nautilus_trader.core.nautilus_pyo3 import Equity
 from nautilus_trader.core.nautilus_pyo3 import FuturesContract
 from nautilus_trader.core.nautilus_pyo3 import FuturesSpread
+from nautilus_trader.core.nautilus_pyo3 import IndexInstrument
 from nautilus_trader.core.nautilus_pyo3 import InstrumentId
 from nautilus_trader.core.nautilus_pyo3 import Money
 from nautilus_trader.core.nautilus_pyo3 import OptionContract
@@ -510,6 +513,57 @@ class TestInstrumentProviderPyo3:
             min_quantity=None,
             max_price=None,
             min_price=None,
+            ts_event=0,
+            ts_init=0,
+        )
+
+    @staticmethod
+    def commodity() -> Commodity:
+        return Commodity(
+            instrument_id=InstrumentId(symbol=Symbol("CL"), venue=Venue("NYMEX")),
+            raw_symbol=Symbol("CL"),
+            asset_class=AssetClass.COMMODITY,
+            quote_currency=_USD,
+            price_precision=2,
+            size_precision=0,
+            price_increment=Price.from_str("0.01"),
+            size_increment=Quantity.from_int(1),
+            lot_size=Quantity.from_int(1),
+            ts_event=0,
+            ts_init=0,
+        )
+
+    @staticmethod
+    def index_instrument() -> IndexInstrument:
+        return IndexInstrument(
+            instrument_id=InstrumentId(symbol=Symbol("SPX"), venue=Venue("INDEX")),
+            raw_symbol=Symbol("SPX"),
+            currency=_USD,
+            price_precision=2,
+            size_precision=0,
+            price_increment=Price.from_str("0.01"),
+            size_increment=Quantity.from_int(1),
+            ts_event=0,
+            ts_init=0,
+        )
+
+    @staticmethod
+    def cfd() -> Cfd:
+        return Cfd(
+            instrument_id=InstrumentId.from_str("AUDUSD.OANDA"),
+            raw_symbol=Symbol("AUD/USD"),
+            asset_class=AssetClass.FX,
+            quote_currency=_USD,
+            base_currency=Currency.from_str("AUD"),
+            price_precision=5,
+            size_precision=0,
+            price_increment=Price.from_str("0.00001"),
+            size_increment=Quantity.from_int(1),
+            lot_size=Quantity.from_int(1000),
+            margin_init=Decimal("0.03"),
+            margin_maint=Decimal("0.03"),
+            maker_fee=Decimal("0.00002"),
+            taker_fee=Decimal("0.00002"),
             ts_event=0,
             ts_init=0,
         )